+++ CHANGELOG.md
@@ -10,6 +10,10 @@
   are required. If the field is null, REQUIRED and UNIQUE_IDENTIFIER, a UUID is
   automatically generated. If the field is null and REQUIRED, an Exception is thrown.
 
+* Add QueryOption.EXPAND_LINKS for expanding selected link values with the
+  target document state and including it in the query results. The serialized
+  state, is placed in the results.selectedLinks map.
+
 * Add PropertyUsageOption.SENSITIVE to describe fields that contain sensitive
   information. When marked, the field will be hidden when serializing to JSON
   using toJson(boolean hideSensitiveFields, boolean useHtmlFormatting).